feat: update joining security policy
If user owner of the file, or already joined with same gmail, nothing would happed.

If user was joined before with different gmail. Old permission would be revoked, and new one would be created for new gmail and same role as in the old permission.

If user try to join with gmail someone else used, error would be raised